added Sakai 2.9.x ScreenSteps help
This version of the Sakai help was contributed by Tufts University and
has been modified for Trunk (their local branding of Sakai 2.9.x). This
is the ScreenSteps package of the help content. Due to size limitations,
it has been zipped and split into two parts. Please use 7zip to
recombine the file before extracting and importing the package into a
ScreenSteps account.